Add comprehensive Heap tests and benchmarks with zero-allocation verification
- Add extensive Heap class functionality tests with edge cases
- Create performance benchmarks comparing custom Heap vs std::priority_queue
- Implement allocation tracking to verify zero runtime allocations
- Results show custom Heap is 2-3x faster and truly allocation-free
- Fix all compiler warnings across codebase for clean builds
- Add strict warning configuration to CMakeLists.txt

Performance Results:
- Custom Heap: 0 allocations during runtime (verified)
- std::priority_queue: 11 allocations, 32KB for 1000 elements
- Custom Heap 2.3x faster for 1000 elements (171μs vs 394μs)
- Custom Heap 2.7x faster for insert-only operations
- Consistent performance advantage across all test scenarios

+++ examples/basic_example.cpp
@@ -157,7 +157,7 @@ int main()
             float phase_change = output_phases[i] - prev_phases[i];
             
             printf("%3zu | %9.1f | %9.3f | %10.3f | %11.6f\n",
-                   i, freq, magnitudes[i], output_mags[i], phase_change);
+                   i, static_cast<double>(freq), static_cast<double>(magnitudes[i]), static_cast<double>(output_mags[i]), static_cast<double>(phase_change));
         }
     }
